fix: Level 1 — correctness & memory safety
Memory:
- ctp_free_dataset: free db_search and plotProperties (were leaked); make
  it NULL-safe so it can clean up a partially-initialized dataset
- ctp_initialize_dataset: calloc + goto-fail cleanup; null-check before use
- ctp_analyze_mean/sd/md: free the internal n (and mean) arrays
- ctp_analyze_*_search: drop the calloc that was immediately overwritten
- ctp_plot_analyze: free each inner allocation

Correctness:
- ctp_utils_normalizes: fix out-of-bounds write initialDone[2] -> [0]
  (and the early-break it disabled); guard divide-by-zero on constant columns
- ctp_analyze_sd: take sqrt (was returning variance)
- ctp_analyze_md: use fabs, not integer abs (was truncating to 0)
- ctp_analyze_md_search: call md, not sd
- ctp_findMany: add 'ne', fix broken 'gte', reject unknown operators;
  collapse 5 duplicated branches into one comparison
- ctp_set_grap_point_*: write to char[] buffers, not string literals (UB)
- ctp_plot: honor the scatter_plot flag (was checking line_plot)
- fix mispasted error messages in add_row/add_data/add_label
- main.c: %u for sizeof

Add tests/test_ctp.c (asserts mean/sd/md + search counts) and golden
output snapshots; examples render byte-identical to baseline.

+++ src/main.c
@@ -41,7 +41,7 @@ int main()
     ctp_plot_analyze(dataSet, mean);
     ctp_plot(dataSet);
 
-    printf("%d %d %d %d", sizeof(char), sizeof(int), sizeof(float), sizeof(double));
+    printf("%u %u %u %u", (unsigned)sizeof(char), (unsigned)sizeof(int), (unsigned)sizeof(float), (unsigned)sizeof(double));
 
     // ctp_findMany(dataSet, 0, "gte", -1);
     // ctp_findMany(dataSet, 0, "lte", 2);
